+++ lib/methods/files.py
@@ -24,10 +24,10 @@ class FilesMethod(BaseMethod):
 
   @staticmethod
   def applyConfig(config, settings):
-    if config['filesFolder'].find(config['rootFolder']) < 0:
-      config['filesFolder'] = config['rootFolder'] + config['filesFolder']
-    if config['siteFolder'].find(config['rootFolder']) < 0:
-      config['siteFolder'] = config['rootFolder'] + config['siteFolder']
+    keys = ['filesFolder', 'siteFolder']
+    for key in keys:
+      if key in config and config[key].find(config['rootFolder']) < 0:
+        config[key] = config['rootFolder'] + config[key]
 
     BaseMethod.addExecutables(config, ['tar', 'rsync'])
 
@@ -35,9 +35,8 @@ class FilesMethod(BaseMethod):
   def tarFiles(self, config, filename, source_folders, type):
     excludeFiles = configuration.getSettings('excludeFiles')
     excludeFiles = excludeFiles[type] if type in excludeFiles else False
-    cmd = 'tar'
+    cmd = '#!tar'
     if excludeFiles:
-      print excludeFiles
       cmd += ' --exclude="'  + '" --exclude="'.join(excludeFiles) + '"'
     cmd += ' -czPf ' + filename
     cmd += ' ' + ' '.join(source_folders)
@@ -50,6 +49,13 @@ class FilesMethod(BaseMethod):
 
     baseName = kwargs['baseName']
     filename = config['backupFolder'] + "/" + '--'.join(baseName) + ".tgz"
+
+    self.backupFiles(config, backup_file_name=filename)
+
+
+  def backupFiles(self, config, **kwargs):
+    self.setRunLocally(config)
+    filename = kwargs['backup_file_name']
     source_folders = kwargs['sourceFolders'] if 'sourceFolders' in kwargs else []
 
     if 'filesFolder' in config:
@@ -60,6 +66,7 @@ class FilesMethod(BaseMethod):
     if len(source_folders) > 0:
       self.tarFiles(config, filename, source_folders, 'backup')
       print green('Files dumped into "%s"' % filename)
+
 
   def listBackups(self, config, results, **kwargs):
     files = self.list_remote_files(config['backupFolder'], ['*.tgz'])